refactor links into AddressUtils, move to GeneralInfo
